move template codes from stanza to lib
