+++ lib/ToolTip.js
@@ -1,5 +1,6 @@
 import { LitElement, css, html } from "lit";
 import { createRef, ref } from "lit/directives/ref.js";
+import Handlebars from "handlebars/dist/handlebars.js";
 
 /**
  * ノードリスト内に特定のノードが存在するか確認するユーティリティ関数
@@ -21,6 +22,17 @@ function isInNodeList(nodeList, node) {
  * LitElement を拡張して、ツールチップ表示の動作を実装
  */
 export default class ToolTip extends LitElement {
+  // ツールチップの対象となる複数のノードを保持する配列
+  nodes = [];
+  // ツールチップ表示状態のフラグ
+  showing = false;
+  // ツールチップに表示するテキストやHTMLコンテンツ
+  tooltipContent = "";
+  // ツールチップコンテンツをHTMLとして解釈するためのヘルパー関数
+  tooltipHTML = (content) => html`${content}`;
+  handlebarsTemplate = null;
+  tooltipsVariables = null;
+
   constructor() {
     super();
 
@@ -28,15 +40,6 @@ export default class ToolTip extends LitElement {
     this.tooltip = createRef();
     // 起点要素への参照（相対位置計算のため）
     this.origin = createRef();
-
-    // ツールチップの対象となる複数のノードを保持する配列
-    this.nodes = [];
-    // ツールチップ表示状態のフラグ
-    this.showing = false;
-    // ツールチップに表示するテキストやHTMLコンテンツ
-    this.tooltipContent = "";
-    // ツールチップコンテンツをHTMLとして解釈するためのヘルパー関数
-    this.tooltipHTML = (content) => html`${content}`;
 
     // マウスアウトイベント用のリスナー
     this.mouseEL = (e) => {
@@ -155,6 +158,15 @@ export default class ToolTip extends LitElement {
     super.disconnectedCallback();
   }
 
+  setTemplate(templateStr) {
+    // Handlebars テンプレート文を生成
+    this.handlebarsTemplate = Handlebars.compile(templateStr);
+    const tooltipsVariables = this.getTemplateVariables(templateStr);
+    console.log(tooltipsVariables);
+  }
+
+
+
   /**
    * 対象ノードの設定を行うメソッド
    * 親要素にマウスアウトのイベントリスナーを設定する
@@ -164,6 +176,65 @@ export default class ToolTip extends LitElement {
     this.nodes = nodes;
     this.parentElement.addEventListener("mouseout", this.mouseEL);
   }
+
+  /**
+   * ツールチップの内容を更新するメソッド
+   * @param {object} contents - ツールチップに表示するコンテンツ
+   */
+  updateTooltipContents(contents) {
+    // Handlebars テンプレートを使用してコンテンツを生成
+    if (this.handlebarsTemplate) {
+      return this.handlebarsTemplate(contents);
+    } else {
+      return contents;
+    }
+  }
+
+  /**
+ * Handlebars テンプレート文字列から、使われているトップレベルの変数名を抽出する関数
+ * @param {string} templateStr — Handlebars テンプレート文字列
+ * @returns {string[]} — 使われている変数名のユニークな配列
+ */
+  getTemplateVariables(templateStr) {
+    // 1. テンプレート文字列を AST にパース
+    const ast = Handlebars.parse(templateStr);
+
+    // 2. 変数名を格納する Set
+    const vars = new Set();
+
+    // 3. 再帰的に AST をトラバースする関数
+    function walk(node) {
+      if (!node || typeof node !== 'object') { return; }
+
+      // MustacheStatement（{{foo}}）や BlockStatement（{{#if foo}}...{{/if}}）などをチェック
+      if (
+        (node.type === 'MustacheStatement') ||
+        (node.type === 'BlockStatement') ||
+        (node.type === 'PartialStatement') ||
+        (node.type === 'SubExpression')
+      ) {
+        // node.path.parts = ['foo', 'bar', ...] の形式で分割された名前空間
+        if (node.path && Array.isArray(node.path.parts)) {
+          // トップレベルのキー部分だけ取りたいなら parts[0] を使う
+          vars.add(node.path.parts[0]);
+        }
+      }
+
+      // 4. 子ノードにも再帰的に.walk
+      for (const key of Object.keys(node)) {
+        const child = node[key];
+        if (Array.isArray(child)) {
+          child.forEach(walk);
+        } else {
+          walk(child);
+        }
+      }
+    }
+
+    walk(ast);
+    return Array.from(vars);
+  }
+
 }
 
 // カスタム要素として登録
